requirements.txt
torch>=2.4.0+cu124
torchvision>=0.19.0+cu124
datasets>=2.20.0
transformers>=4.43.3
tqdm>=4.66.4
numpy>=1.26.4
PyYAML>=6.0.1